Integrate remote display providers into Quick Settings.
This is a first pass at integrating new remote display functionality
into Quick Settings.  The Wireless Display card which previously
only supported Wifi Display is reimplemented to use the media router
to enumerate available remote display routes.  This ensures that the
user is presented with a consistent state regarding the currently
selected display route.

In this patch, the Wireless Display card still launches the old
Settings preference page for Wifi Display when clicked.  This will be
addressed in future patches.

As part of this change, it was necessary to derive some new lifecycle
information regarding the visibility of the Quick Settings model.
When Quick Settings is shown, an onPrepare event is delivered to
give the model a chance to update its state.  Likewise when Quick
Settings is hidden, an onUnprepare event is delivered.

These events allow the system to determine precisely when remote
display discovery is required to update the UI so as not to waste
power performing discovery in the background all of the time.

Bug: 11257292
Change-Id: Id802aa0983b625aeb972b5d123e4cc080dd6705f

Conflicts:
	packages/SystemUI/src/com/android/systemui/statusbar/phone/QuickSettings.java
	packages/SystemUI/src/com/android/systemui/statusbar/phone/QuickSettingsModel.java
	packages/SystemUI/src/com/android/systemui/statusbar/phone/QuickSettingsTileView.java