Make quick settings show the standard media router icon.
Fixed a bug in ImageView where we failed to inform a newly updated
Drawable about the visibility state.  This caused AnimationDrawables
to not animate when attached to an existing ImageView *unless* that
ImageView happened to be attached to the window *later* or have
its visibility toggled for some other reason.

Bug: 11257292
Change-Id: Iba9e0db5ba0db2b022950aec0c6f60a435da8ad2

Conflicts:
	core/res/res/drawable-hdpi/ic_lock_screenshot.png