Battery Text: Fix rounding issues and take density into account
We are a Multi-Dpi ROM... on 4.2 and 4.3 the text in the circle battery icon
was a small bit up on a lower dpi. The original author compensated this with
just a +1 to the Y value. Thats not the right way...
Calculate the current dpi into the offset and all is well.

We want a perfect 4.4...so this nitpicks even if we are talking about
0.5dp should be fixed

Change-Id: I5c6c1e237cd9049101ee88347a9b4d4b5a2b9682