Base: Allow hiding notification icons
When long pressing a notification, add another menu
item which allows the user to hide the notification icon
(but keep the actual notification in the dropdown).

This is meant for applications which have services running
in the statusbar which the user is aware of, so right now
we hide the icon by the package.

Change-Id: Id319e1b14d0cf8957cc718fcc27bf0dcb394bc8f

Conflicts:
	packages/SystemUI/res/menu/notification_popup_menu.xml
	packages/SystemUI/res/values/rf_strings.xml
	packages/SystemUI/src/com/android/systemui/statusbar/BaseStatusBar.java