SystemUI: UX: Make non-flip tiles slightly flippable
I always end up with people trying to flip non-flip tiles, and thus
it just acts like a long press on that tile. To give a better feedback
on tiles that aren't flippables, we let the tile slightly flip when
dragged, then get back into place when the finger goes up.

Change-Id: I14fddfc2033320bf447aab0bc6a1a49f77d8555e

Conflicts:
	packages/SystemUI/src/com/android/systemui/statusbar/phone/QuickSettingsTileView.java